Make edit description pure
