+++ src/index.js
@@ -1,6 +1,6 @@
 import './style.css';
 import { display } from './Module/ui.js';
 import './Module/user.js';
-import './Module/state.js';
+import './Module/status.js';
 
 document.addEventListener('DOMContentLoaded', display);
